Decoding cleanups (#1321)
* Decoding cleanups

* Split extractDef

* Extract struct types to string mapping
