+++ packages/types/src/codec/AbstractInt.ts
@@ -45,7 +45,7 @@ export default abstract class AbstractInt extends BN implements Codec {
     // This function returns a string, which will be passed in the BN
     // constructor. It would be ideal to actually return a BN, but there's a
     // bug: https://github.com/indutny/bn.js/issues/206.
-    if (isHex(value)) {
+    if (isHex(value, -1, true)) {
       return hexToBn(value, { isLe: false, isNegative }).toString();
     } else if (isU8a(value)) {
       return AbstractInt.decodeAbstracIntU8a(value, bitLength, isNegative);
